+++ transport/wseb/src/main/java/org/kaazing/gateway/transport/wseb/WsebAcceptProcessor.java
@@ -250,7 +250,7 @@ public class WsebAcceptProcessor extends BridgeAcceptProcessor<WsebSession> {
         // check to see if we need to add a padding message to the end of sent messages
         long writtenBytes = session.getWrittenBytes();
         Long bytesWrittenOnLastFlush = (Long)session.getAttribute(WsebAcceptor.BYTES_WRITTEN_ON_LAST_FLUSH_KEY);
-        if (bytesWrittenOnLastFlush == null || writtenBytes != bytesWrittenOnLastFlush.longValue()) {
+        if (bytesWrittenOnLastFlush == null || writtenBytes != bytesWrittenOnLastFlush) {
             // Block Padding is required
             session.write(WsebFrameEncoder.BLOCK_PADDING_MESSAGE);
             session.setAttribute(WsebAcceptor.BYTES_WRITTEN_ON_LAST_FLUSH_KEY, writtenBytes + 4096);
